Added Decor & Mediation error
Followed prompts to fix the mediation error. The prompts instructed me to link the project to myself as the active organization. Not sure if that affects how Unity allows access to the game. Also added some minor decorations. Added more books and pillows to the environment, threw in 2 cats, and created 2 materials for the pillows.